# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Finance App API",
    "version": "0.1.0",
    "description": "Personal finance management platform.\n\nThis document is **generated from the running Express app** by `npm run generate:openapi`, so the paths, parameters and request bodies below are the ones the server actually enforces. It is committed to the repository and checked in CI, which is what stops it from drifting.\n\nMoney crosses this API as a **decimal string**, never a JSON number. Every amount is stored as `NUMERIC(19,4)`.\n\nResponse bodies are not described yet — that is phase 2."
  },
  "servers": [
    {
      "url": "http://localhost:4000"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "A short-lived access token from `POST /api/v1/auth/login` or `/auth/refresh`. Refresh tokens are opaque, rotate on every use, and are tracked in families."
      }
    },
...
    "responses": {
      "BadRequest": {
        "description": "The request was malformed — unparseable JSON, or a value the database refused outright.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "No bearer token was sent, or it has expired or belongs to an inactive account.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "The caller is authenticated but lacks the required role in this workspace.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "No such resource, or none the caller is allowed to know exists.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "ValidationFailed": {
        "description": "One or more fields were rejected. `error.details` names each one, already translated into the language resolved for the request.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "RateLimited": {
        "description": "Too many requests; the caller has exceeded the rate limit for this window.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    }
  },
  "security": [],
  "paths": {
...
    "/api/v1/workspaces/{workspaceId}/transactions": {
      "get": {
        "operationId": "getWorkspacesByWorkspaceIdTransactions",
        "tags": [
          "transactions"
        ],
        "description": "Requires the `viewer` role or higher in the workspace.",
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "default": 1,
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "default": 50,
              "type": "integer",
              "minimum": 1,
              "maximum": 200
            }
          },
          {
            "name": "accountIds",
            "in": "query",
            "required": false,
            "description": "UUIDs, either comma-separated in one parameter or repeated across several. Every entry must be a UUID.",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              ]
            }
          },
          {
            "name": "categoryIds",
            "in": "query",
            "required": false,
            "description": "UUIDs, either comma-separated in one parameter or repeated across several. Every entry must be a UUID.",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              ]
            }
          },
          {
            "name": "tagIds",
            "in": "query",
            "required": false,
            "description": "UUIDs, either comma-separated in one parameter or repeated across several. Every entry must be a UUID.",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              ]
            }
          },
          {
            "name": "includeSubcategories",
            "in": "query",
            "required": false,
            "description": "A boolean written into a query string. `true`, `1` and `yes` are true; anything else listed is false.",
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "string",
                  "enum": [
                    "0",
                    "1",
                    "true",
                    "false",
                    "yes",
                    "no"
                  ]
                }
              ]
            }
          },
          {
            "name": "types",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              ]
            }
          },
          {
            "name": "statuses",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              ]
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "description": "A calendar date as `YYYY-MM-DD`. The date must exist: `2025-02-30` is rejected.",
            "schema": {
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "description": "A calendar date as `YYYY-MM-DD`. The date must exist: `2025-02-30` is rejected.",
            "schema": {
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
              "type": "string"
            }
          },
          {
            "name": "minAmount",
            "in": "query",
            "required": false,
            "description": "A decimal amount with at most 15 integer digits and 4 decimal places, matching the `NUMERIC(19,4)` column it is stored in. Send it as a string: a JSON number is accepted for convenience but loses precision on the way in.",
            "schema": {
              "pattern": "^-?\\d{1,15}(\\.\\d{1,4})?$",
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "number"
                }
              ]
            }
          },
          {
            "name": "maxAmount",
            "in": "query",
            "required": false,
            "description": "A decimal amount with at most 15 integer digits and 4 decimal places, matching the `NUMERIC(19,4)` column it is stored in. Send it as a string: a JSON number is accepted for convenience but loses precision on the way in.",
            "schema": {
              "pattern": "^-?\\d{1,15}(\\.\\d{1,4})?$",
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "number"
                }
              ]
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 200
            }
          },
          {
            "name": "createdBy",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
            }
          },
          {
            "name": "isReconciled",
            "in": "query",
            "required": false,
            "description": "A boolean written into a query string. `true`, `1` and `yes` are true; anything else listed is false.",
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "string",
                  "enum": [
                    "0",
                    "1",
                    "true",
                    "false",
                    "yes",
                    "no"
                  ]
                }
              ]
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "schema": {
              "default": "occurredOn",
              "type": "string",
              "enum": [
                "occurredOn",
                "amount",
                "createdAt"
              ]
            }
          },
          {
            "name": "sortDirection",
            "in": "query",
            "required": false,
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "2XX": {
            "description": "The request succeeded. Response bodies are not described yet — see the note on phase 2 in `docs/decisions.md` — so a caller should read `docs/api.md` for the shape."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "operationId": "postWorkspacesByWorkspaceIdTransactions",
        "tags": [
          "transactions"
        ],
        "description": "Requires the `editor` role or higher in the workspace.",
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "accountId": {
                    "type": "string",
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
                  },
                  "categoryId": {
                    "anyOf": [
                      {
                        "type": "string",
                        "format": "uuid",
                        "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "income",
                      "expense"
                    ]
                  },
                  "amount": {
                    "pattern": "^-?\\d{1,15}(\\.\\d{1,4})?$",
                    "description": "A decimal amount greater than zero. \"Greater than zero\" cannot be expressed against a decimal held as a string, so it is stated here rather than published as a constraint.",
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "number"
                      }
                    ]
                  },
                  "description": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "merchant": {
                    "anyOf": [
                      {
                        "type": "string",
                        "maxLength": 120
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "notes": {
                    "anyOf": [
                      {
                        "type": "string",
                        "maxLength": 2000
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "occurredOn": {
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
                    "description": "A calendar date as `YYYY-MM-DD`. The date must exist: `2025-02-30` is rejected.",
                    "type": "string"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "cleared",
                      "pending",
                      "scheduled"
                    ]
                  },
                  "paidByUserId": {
                    "anyOf": [
                      {
                        "type": "string",
                        "format": "uuid",
                        "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "tagIds": {
                    "maxItems": 20,
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
                    }
                  }
                },
                "required": [
                  "accountId",
                  "type",
                  "amount",
                  "description",
                  "occurredOn"
                ]
              }
            }
          }
        },
        "responses": {
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "2XX": {
            "description": "The request succeeded. Response bodies are not described yet — see the note on phase 2 in `docs/decisions.md` — so a caller should read `docs/api.md` for the shape."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/workspaces/{workspaceId}/transactions/transfers": {
      "post": {
        "operationId": "postWorkspacesByWorkspaceIdTransactionsTransfers",
        "tags": [
          "transactions"
        ],
        "description": "Requires the `editor` role or higher in the workspace.",
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fromAccountId": {
                    "type": "string",
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
                  },
                  "toAccountId": {
                    "type": "string",
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
                  },
                  "amount": {
                    "pattern": "^-?\\d{1,15}(\\.\\d{1,4})?$",
                    "description": "A decimal amount greater than zero. \"Greater than zero\" cannot be expressed against a decimal held as a string, so it is stated here rather than published as a constraint.",
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "number"
                      }
                    ]
                  },
                  "destinationAmount": {
                    "pattern": "^-?\\d{1,15}(\\.\\d{1,4})?$",
                    "description": "A decimal amount greater than zero. \"Greater than zero\" cannot be expressed against a decimal held as a string, so it is stated here rather than published as a constraint.",
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "number"
                      }
                    ]
                  },
                  "description": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "notes": {
                    "anyOf": [
                      {
                        "type": "string",
                        "maxLength": 2000
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "occurredOn": {
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
                    "description": "A calendar date as `YYYY-MM-DD`. The date must exist: `2025-02-30` is rejected.",
                    "type": "string"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "cleared",
                      "pending"
                    ]
                  }
                },
                "required": [
                  "fromAccountId",
                  "toAccountId",
                  "amount",
                  "description",
                  "occurredOn"
                ]
              }
            }
          }
        },
        "responses": {
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "2XX": {
            "description": "The request succeeded. Response bodies are not described yet — see the note on phase 2 in `docs/decisions.md` — so a caller should read `docs/api.md` for the shape."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/workspaces/{workspaceId}/transactions/bulk-categorize": {
      "post": {
        "operationId": "postWorkspacesByWorkspaceIdTransactionsBulkCategorize",
        "tags": [
          "transactions"
        ],
        "description": "Requires the `editor` role or higher in the workspace.",
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "transactionIds": {
                    "minItems": 1,
                    "maxItems": 500,
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
                    }
                  },
                  "categoryId": {
                    "anyOf": [
                      {
                        "type": "string",
                        "format": "uuid",
                        "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  }
                },
                "required": [
                  "transactionIds",
                  "categoryId"
                ]
              }
            }
          }
        },
        "responses": {
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "2XX": {
            "description": "The request succeeded. Response bodies are not described yet — see the note on phase 2 in `docs/decisions.md` — so a caller should read `docs/api.md` for the shape."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  }
}